Tweaks but still no luck with decoder matching results
Adding the ability to include time lags when computing outputs.
Hardcoding the number of time lags to be 10, per the paper. Also using
the least squares estimation mentioned in a previous paper (JMC 2003),
which should probably be equivalent to normal linear regression. The
residuals are still pretty high, on the order of the actual velocities.

Reading through the paper, the output of the matrix multiplication (i.e.
Y(t) in the paper) isn't cursor velocity; it seems to be "joint"
parameters for elbow and shoulder, which are then put together to get a
virtual cursor position/velocity.